% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate_pages.R
\name{activate_pages}
\alias{activate_pages}
\title{Title}
\usage{
activate_pages(account_name, repo_name, branch)
}
\arguments{
\item{account_name}{Name of the account}

\item{repo_name}{Name of the repo}

\item{branch}{Name of the branch}
}
\value{
activate GitHub pages with correct parameters
}
\description{
Title
}
