% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_host_blogdown.R
\name{github_host_blogdown}
\alias{github_host_blogdown}
\title{github_host_blogdown}
\usage{
github_host_blogdown()
}
\description{
Set blogdown parameters useful for hosting a webiste on GitHub Pages,
}
