% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_github_blogdown.R
\name{github_Rprofile_and_activate_pages}
\alias{github_Rprofile_and_activate_pages}
\title{Title}
\usage{
github_Rprofile_and_activate_pages(account_name, repo_name)
}
\arguments{
\item{account_name}{Name of the branch}

\item{repo_name}{Name of the repo}
}
\value{
Wrapper function
}
\description{
Title
}
