% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_hugo_Rprofile_version.R
\name{install_hugo_Rprofile_version}
\alias{install_hugo_Rprofile_version}
\title{install_hugo_Rprofile_version: install Hugo version found in .Rprofile of the project}
\usage{
install_hugo_Rprofile_version()
}
\description{
Search all versions of Hugo and compare them to recommended version set in the .Rprofile file found in the working directory of the project
If the correct Hugo version is already installed then do nothing
}
