% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebuild_and_serve.R
\name{rebuild_and_serve}
\alias{rebuild_and_serve}
\title{Rebuild_and_serve}
\usage{
rebuild_and_serve(...)
}
\arguments{
\item{...}{Arguments passed to blogdown::build_site() (arguments build_site, run_hugo, and ...).}
}
\description{
Wrapper for functions stop_site(), build_site() and serve_site().
Needed in case you have permission access issues in building .Rmd when site is served.
}
