% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_gh_pages_and_Rprofile.R
\name{set_gh_pages_and_Rprofile}
\alias{set_gh_pages_and_Rprofile}
\title{set_gh_pages_and_Rprofile}
\usage{
set_gh_pages_and_Rprofile(account_name, repo_name)
}
\arguments{
\item{account_name}{Name of the account associated to the repository}

\item{repo_name}{Name of the repository}
}
\description{
Wrapper for functions set_publishDir() and set_github_pages()
Needed in github_host_blogdown.
set_github_pages branch is set to the currently active one.
}
