% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_github_pages.R
\name{set_github_pages}
\alias{set_github_pages}
\title{set_github_pages: activate Github Pages functionality from R}
\usage{
set_github_pages(account, repository, branch = gert::git_branch())
}
\arguments{
\item{account}{Name of the account}

\item{repository}{Name of the repo}

\item{branch}{Name of the branch}
}
\description{
Activate the option 'Pages' of Github.
Account, repository and branch are necessary.
The published folder will be named 'docs'.
}
