% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_publishDir.R
\name{set_publishDir}
\alias{set_publishDir}
\title{set_publishDir: set the blogdown publishing directory to 'docs'}
\usage{
set_publishDir()
}
\description{
Set the parameter publishDir in the project .Rprofile to 'docs'.
If the parameter doesn't already exist it will be create otherwise the old one will be overwritten
}
