% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_all_and_commit.R
\name{stage_all_and_commit}
\alias{stage_all_and_commit}
\title{Stage_all_and_commit}
\usage{
stage_all_and_commit(repo = getwd())
}
\arguments{
\item{repo}{path where the git repository is located. Default to the working directory.}
}
\description{
In some case staging many or big files from rstudio cause git to crash.
This function stage every possible files and open the Rstudio commit window.
}
